+++ CHANGELOG.md
@@ -15,6 +15,8 @@
 - Bug fix: column name completion function does not generate errors when no pgmacstbl is present
   (which can occur when displaying an empty result set).
 
+- Bug fix: updating the last row of a table no longer triggers an error related to row colors.
+
 
 ## [0.16] - 2024-10-27